changed integer to decimal

+++ db/migrate/20160715191235_create_concerts.rb
@@ -13,8 +13,8 @@ class CreateConcerts < ActiveRecord::Migration[5.0]
       t.string :artist3
       t.string :time
       t.string :venue_address
-      t.integer :venue_latitude
-      t.integer :venue_longitude
+      t.decimal :venue_latitude
+      t.decimal :venue_longitude
       t.timestamps
     end
   end
